Add autoloading and namespaces
