+++ includes/class-solo-api-woocommerce-integration-i18n.php
@@ -2,25 +2,22 @@
 /**
  * Define the internationalization functionality
  *
- * Loads and defines the internationalization files for this plugin
- * so that it is ready for translation.
- *
  * @link       https://madebydenis.com
  * @since      1.0.0
  *
- * @package    Solo_Api_Woocommerce_Integration
- * @subpackage Solo_Api_Woocommerce_Integration/includes
+ * @package    Solo_Api_Woocommerce_Integration\Includes
  */
 
+namespace Solo_Api_Woocommerce_Integration\Includes;
+
 /**
- * Define the internationalization functionality.
+ * Define the internationalization functionality class.
  *
  * Loads and defines the internationalization files for this plugin
  * so that it is ready for translation.
  *
  * @since      1.0.0
- * @package    Solo_Api_Woocommerce_Integration
- * @subpackage Solo_Api_Woocommerce_Integration/includes
+ * @package    Solo_Api_Woocommerce_Integration\Includes
  * @author     Denis Žoljom <denis.zoljom@gmail.com>
  */
 class Solo_Api_Woocommerce_Integration_i18n {
