+++ includes/class-solo-api-woocommerce-integration-loader.php
@@ -5,9 +5,10 @@
  * @link       https://madebydenis.com
  * @since      1.0.0
  *
- * @package    Solo_Api_Woocommerce_Integration
- * @subpackage Solo_Api_Woocommerce_Integration/includes
+ * @package    Solo_Api_Woocommerce_Integration\Includes
  */
+
+namespace Solo_Api_Woocommerce_Integration\Includes;
 
 /**
  * Register all actions and filters for the plugin.
@@ -16,8 +17,7 @@
  * the plugin, and register them with the WordPress API. Call the
  * run function to execute the list of actions and filters.
  *
- * @package    Solo_Api_Woocommerce_Integration
- * @subpackage Solo_Api_Woocommerce_Integration/includes
+ * @package    Solo_Api_Woocommerce_Integration\Includes
  * @author     Denis Žoljom <denis.zoljom@gmail.com>
  */
 class Solo_Api_Woocommerce_Integration_Loader {
@@ -56,11 +56,11 @@ class Solo_Api_Woocommerce_Integration_Loader {
    * Add a new action to the collection to be registered with WordPress.
    *
    * @since    1.0.0
-   * @param    string               $hook             The name of the WordPress action that is being registered.
-   * @param    object               $component        A reference to the instance of the object on which the action is defined.
-   * @param    string               $callback         The name of the function definition on the $component.
-   * @param    int                  $priority         Optional. The priority at which the function should be fired. Default is 10.
-   * @param    int                  $accepted_args    Optional. The number of arguments that should be passed to the $callback. Default is 1.
+   * @param    string $hook             The name of the WordPress action that is being registered.
+   * @param    object $component        A reference to the instance of the object on which the action is defined.
+   * @param    string $callback         The name of the function definition on the $component.
+   * @param    int    $priority         Optional. The priority at which the function should be fired. Default is 10.
+   * @param    int    $accepted_args    Optional. The number of arguments that should be passed to the $callback. Default is 1.
    */
   public function add_action( $hook, $component, $callback, $priority = 10, $accepted_args = 1 ) {
     $this->actions = $this->add( $this->actions, $hook, $component, $callback, $priority, $accepted_args );
@@ -70,11 +70,11 @@ class Solo_Api_Woocommerce_Integration_Loader {
    * Add a new filter to the collection to be registered with WordPress.
    *
    * @since    1.0.0
-   * @param    string               $hook             The name of the WordPress filter that is being registered.
-   * @param    object               $component        A reference to the instance of the object on which the filter is defined.
-   * @param    string               $callback         The name of the function definition on the $component.
-   * @param    int                  $priority         Optional. The priority at which the function should be fired. Default is 10.
-   * @param    int                  $accepted_args    Optional. The number of arguments that should be passed to the $callback. Default is 1
+   * @param    string $hook             The name of the WordPress filter that is being registered.
+   * @param    object $component        A reference to the instance of the object on which the filter is defined.
+   * @param    string $callback         The name of the function definition on the $component.
+   * @param    int    $priority         Optional. The priority at which the function should be fired. Default is 10.
+   * @param    int    $accepted_args    Optional. The number of arguments that should be passed to the $callback. Default is 1.
    */
   public function add_filter( $hook, $component, $callback, $priority = 10, $accepted_args = 1 ) {
     $this->filters = $this->add( $this->filters, $hook, $component, $callback, $priority, $accepted_args );
@@ -86,26 +86,25 @@ class Solo_Api_Woocommerce_Integration_Loader {
    *
    * @since    1.0.0
    * @access   private
-   * @param    array                $hooks            The collection of hooks that is being registered (that is, actions or filters).
-   * @param    string               $hook             The name of the WordPress filter that is being registered.
-   * @param    object               $component        A reference to the instance of the object on which the filter is defined.
-   * @param    string               $callback         The name of the function definition on the $component.
-   * @param    int                  $priority         The priority at which the function should be fired.
-   * @param    int                  $accepted_args    The number of arguments that should be passed to the $callback.
-   * @return   array                                  The collection of actions and filters registered with WordPress.
+   * @param    array  $hooks            The collection of hooks that is being registered (that is, actions or filters).
+   * @param    string $hook             The name of the WordPress filter that is being registered.
+   * @param    object $component        A reference to the instance of the object on which the filter is defined.
+   * @param    string $callback         The name of the function definition on the $component.
+   * @param    int    $priority         The priority at which the function should be fired.
+   * @param    int    $accepted_args    The number of arguments that should be passed to the $callback.
+   * @return   array                    The collection of actions and filters registered with WordPress.
    */
   private function add( $hooks, $hook, $component, $callback, $priority, $accepted_args ) {
 
     $hooks[] = array(
-      'hook'          => $hook,
-      'component'     => $component,
-      'callback'      => $callback,
-      'priority'      => $priority,
-      'accepted_args' => $accepted_args
+        'hook'          => $hook,
+        'component'     => $component,
+        'callback'      => $callback,
+        'priority'      => $priority,
+        'accepted_args' => $accepted_args,
     );
 
     return $hooks;
-
   }
 
   /**
@@ -114,7 +113,6 @@ class Solo_Api_Woocommerce_Integration_Loader {
    * @since    1.0.0
    */
   public function run() {
-
     foreach ( $this->filters as $hook ) {
       add_filter( $hook['hook'], array( $hook['component'], $hook['callback'] ), $hook['priority'], $hook['accepted_args'] );
     }
@@ -122,7 +120,5 @@ class Solo_Api_Woocommerce_Integration_Loader {
     foreach ( $this->actions as $hook ) {
       add_action( $hook['hook'], array( $hook['component'], $hook['callback'] ), $hook['priority'], $hook['accepted_args'] );
     }
-
   }
-
 }
