+++ solo-api-woocommerce-integration.php
@@ -8,7 +8,7 @@
  * @package           Solo_Api_Woocommerce_Integration
  *
  * Plugin Name:       Solo API WooCommerce Integration
- * Plugin URI:        https://madebydenis/solo-api-woocommerce-integration
+ * Plugin URI:        https://madebydenis.com/solo-api-woocommerce-integration
  * Description:       This plugin provides integration with the SOLO API service with WooCommerce.
  * Version:           1.0.0
  * Author:            Denis Žoljom
@@ -19,12 +19,19 @@
  * Domain Path:       /languages
  */
 
+namespace Solo_Api_Woocommerce_Integration;
+use Solo_Api_Woocommerce_Integration\Includes as Includes;
+
 // If this file is called directly, abort.
 if ( ! defined( 'WPINC' ) ) {
   die;
 }
 
-define( 'PLUGIN_VERSION', '1.0.0' );
+define( 'SAWI_PLUGIN_VERSION', '1.0.0' );
+define( 'SAWI_PLUGIN_NAME', 'solo-api-woocommerce-integration' );
+
+// Include the autoloader so we can dynamically include the rest of the classes.
+include_once( 'lib/autoloader.php' );
 
 /**
  * The code that runs during plugin activation.
@@ -32,16 +39,10 @@ define( 'PLUGIN_VERSION', '1.0.0' );
  */
 function activate_solo_api_woocommerce_integration() {
   require_once plugin_dir_path( __FILE__ ) . 'includes/class-solo-api-woocommerce-integration-activator.php';
-  Solo_Api_Woocommerce_Integration_Activator::activate();
+  Includes\Solo_Api_Woocommerce_Integration_Activator::activate();
 }
 
-register_activation_hook( __FILE__, 'activate_solo_api_woocommerce_integration' );
-
-/**
- * The core plugin class that is used to define internationalization,
- * admin-specific hooks, and public-facing site hooks.
- */
-require plugin_dir_path( __FILE__ ) . 'includes/class-solo-api-woocommerce-integration.php';
+register_activation_hook( __FILE__, __NAMESPACE__ . '\\activate_solo_api_woocommerce_integration' );
 
 /**
  * Begins execution of the plugin.
@@ -54,8 +55,9 @@ require plugin_dir_path( __FILE__ ) . 'includes/class-solo-api-woocommerce-integ
  */
 function run_solo_api_woocommerce_integration() {
 
-  $plugin = new Solo_Api_Woocommerce_Integration();
+  $plugin = new Includes\Solo_Api_Woocommerce_Integration();
   $plugin->run();
 
 }
+
 run_solo_api_woocommerce_integration();
